trading-system/pyproject.toml
[project]
name = "trading-system"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "requests>=2.28",
    "pyyaml>=6.0",
]
